small update for the makefile

+++ srcs/ft_echo.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   echo.c                                             :+:      :+:    :+:   */
+/*   ft_echo.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: earnaud <earnaud@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/30 14:26:42 by earnaud           #+#    #+#             */
-/*   Updated: 2021/03/30 17:06:34 by earnaud          ###   ########.fr       */
+/*   Updated: 2021/03/31 12:17:43 by earnaud          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/builtin.h"
+#include "builtin.h"
 
 int ft_echo(int n, char *str, int fd)
 {
-	int i = 0;
+	size_t i = 0;
 	while (str[i] && (!n || i <= ft_strlen(str)))
 	{
 		write(fd, str + i++, 1);
